+++ libexif/olympus/mnote-olympus-entry.c
@@ -352,7 +352,7 @@ mnote_olympus_entry_get_value (MnoteOlympusEntry *entry, char *v, unsigned int m
 		CC (entry->components, 4, v, maxlen);
 		vl =  exif_get_long (entry->data, entry->order);
 		a = (vl>>24)&0xff; b = (vl>>16)&0xff; c = (vl>>8)&0xff;
-		snprintf (v, maxlen, "%.1f",  c?(float)a*((float)b/(float)c):0 );
+		snprintf (v, maxlen, "%.1f",  c?(double)a*((double)b/(double)c):0 );
 		break;
 	}
 	case MNOTE_NIKON_TAG_FLASHEXPCOMPENSATION:
